Better handling of view timeouts
This should fix problems caused by too infrequently checking for views that have timed out. It used to be possible for the event loop to run for a long time without getting any messages from kafka, and during this time the timeouts were not checked, then the next message that was received could cause a publication of a view newer than others that had actually timed out, thus effectively skipping a view and causing problems for the consumers who need the diff frame.