do not publish a view if earlier views is in queue
When some collectors come back after certain failure state at sync time (00:00, 08:00, and 16:00), it causes the system to believe we have all the collectors for the view and publish it. When doing so the views right before the sync time would be discarded.

This commit added a new checking condition to `maybe_publish_view` procedure, so that if there is some view that the program is tracking and is earlier than the current view, it should not publishing it yet, and wait for the earlier view to be published.